+++ src/pages/ProductList/ProductList.tsx
@@ -1,11 +1,10 @@
 import { useQuery } from '@tanstack/react-query'
-import { AsideFilter } from './AsideFilter'
-import Product from './Product/Product'
-import { SortProductList } from './SortProductList'
+import { AsideFilter } from './components/AsideFilter'
+import Product from './components/Product/Product'
+import { SortProductList } from './components/SortProductList'
 import useQueryParams from 'src/hooks/useQueryParams'
 import productApi from 'src/apis/product.api'
 import { Pagination } from 'src/components/Pagination'
-import { useState } from 'react'
 import { ProductListConfig } from 'src/types/product.type'
 import { omitBy, isUndefined } from 'lodash'
 import categoryApi from 'src/apis/category.api'
