+++ src/pages/ProductList/components/AsideFilter/AsideFilter.tsx
@@ -3,7 +3,7 @@ import { Button } from 'src/components/Button'
 import path from 'src/constants/path'
 import { CategoryIcon, DropDownArrowRightIcon, FilterIcon, StarIcon } from 'src/icons'
 import { Category } from 'src/types/category.type'
-import { QueryConfig } from '../ProductList'
+import { QueryConfig } from '../../ProductList'
 import classNames from 'classnames'
 import { InputNumber } from 'src/components/InputNumber'
 import { useForm, Controller } from 'react-hook-form'
